DRY Practice
Practiced DRY principle which states that you shouldn't be repeating yourself and suggests you to write code that is easy to maintain to reduce redundancy. In this project i simply refactored the working code which was violating the DRY principle and then implemented a simple function that satisfy the needs of all four functions and reduced the code a lot.