Practice: Inheritance
Inheritance is is a concept in which we can reuse a piece of code and also model the data in such a way like child parent hierarchy. In this lesson exercise i learnt how and when to use inheritance it's syntax, super keyword, passing child instance data to parent using inheritance and super keyword and also wrote some useful static methods which can be called directly in the class.