Practice: Polymorphism
Polymorphism generally refers to function overloading and overriding in object oriented programming. function overloading is used when we want to use the same function name multiple times having different parameters. Overriding on the other hand means to inherit a parent class but doesn't want to use that particular method or want some different functionality in the context of the current class when a instance is trying to use it. for example a animal class having sound method and a dog class inherits animal class  . dog class wants to use other method of that class except for the sound in this case dog class can define the sound method in itself. this behaviour overrides the sound method and when called from an instance of dog class javascript first searches sound method in dog class it founds there so no need to search for the sound method in parent class which is animal.