Exercise: Context in Regular Functions
When an unbound regular function is invoked function-style, the context inside of the function will be the global context.
If it is invoked method-style, the context inside of the function will be the method's object. 
If it is invoked constructor-style, the context inside of the function will be the newly created instance of the constructor's class.