Practice: ES5 Class Syntax
In this exercise i created two static and instance methods of Dog and Person class using ES5 class syntax. I basically converted the given ES6 class code and translated it to ES5 and passed all of the test cases. In the array.js file however i used monkeypatching technique to add instance method to a existing javascript built in class Array. The instance method isEqual determines whether or not each element of instance Array is equal to the given array provided by a user.